requirements.txt
# EcoLink Australia — Python Dependencies

# AI / LLM
groq>=0.9.0
google-generativeai>=0.5.0

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary>=2.9.9

# Cache / Queue
redis>=5.0.0

# HTTP client (Xero OAuth, external APIs)
httpx>=0.27.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.9.0

# Utilities
feedparser>=6.0.11
